LayoutTests/imported/w3c/web-platform-tests/intersection-observer/timestamp-expected.txt



PASS Check that timestamps correspond to the to execution context that created the observer. 
FAIL First rAF after iframe is loaded. assert_equals: One notification to top window observer. expected 1 but got 0
FAIL Generate notifications. undefined is not an object (evaluating 'topWindowEntries[1].time')
